+++ src/containers/EmailVerificationPage/EmailVerificationPage.js
@@ -158,9 +158,4 @@ const EmailVerificationPage = compose(
   injectIntl
 )(EmailVerificationPageComponent);
 
-EmailVerificationPage.loadData = (params, search) => {
-  const token = parseVerificationToken(search);
-  return verify(token);
-};
-
 export default EmailVerificationPage;