+++ src/containers/InboxPage/InboxPage.js
@@ -38,7 +38,6 @@ import {
 import { TopbarContainer, NotFoundPage } from '../../containers';
 import config from '../../config';
 
-import { loadData } from './InboxPage.duck';
 import css from './InboxPage.module.css';
 
 const formatDate = (intl, date) => {
@@ -441,7 +440,5 @@ const InboxPage = compose(
   connect(mapStateToProps),
   injectIntl
 )(InboxPageComponent);
-
-InboxPage.loadData = loadData;
 
 export default InboxPage;