+++ src/containers/ListingPage/SectionHeading.js
@@ -45,7 +45,11 @@ const SectionHeading = props => {
           {showContactUser ? (
             <span className={css.contactWrapper}>
               <span className={css.separator}>•</span>
-              <InlineTextButton rootClassName={css.contactLink} onClick={onContactUser}>
+              <InlineTextButton
+                rootClassName={css.contactLink}
+                onClick={onContactUser}
+                enforcePagePreloadFor="SignupPage"
+              >
                 <FormattedMessage id="ListingPage.contactUser" />
               </InlineTextButton>
             </span>