+++ src/containers/StripePayoutPage/StripePayoutPage.js
@@ -27,7 +27,7 @@ import {
 } from '../../components';
 import { StripeConnectAccountForm } from '../../forms';
 import { TopbarContainer } from '..';
-import { savePayoutDetails, loadData } from './StripePayoutPage.duck';
+import { savePayoutDetails } from './StripePayoutPage.duck';
 
 import css from './StripePayoutPage.module.css';
 
@@ -291,7 +291,5 @@ const StripePayoutPage = compose(
   ),
   injectIntl
 )(StripePayoutPageComponent);
-
-StripePayoutPage.loadData = loadData;
 
 export default StripePayoutPage;